requirements.txt
google-cloud-bigquery==3.39.0
google-cloud-secret-manager==2.26.0
google-cloud-logging==3.13.0

requests==2.32.5
python-dateutil==2.9.0.post0
pytz==2025.2
pydantic==2.12.5
pydantic-settings==2.12.0
tenacity==9.1.2

google-auth==2.45.0
python-dotenv==1.2.1

# For min_winning_bid tool
pandas==2.2.3
